+++ app/services/task_list_creator.rb
@@ -15,7 +15,14 @@ class TaskListCreator
       section = Section.create(title: workflow_section.fetch("title"), order: index, project: project)
 
       workflow_section.fetch("tasks").each_with_index do |workflow_task, index|
-        task = Task.create(title: workflow_task.fetch("title"), order: index, section: section)
+        task = Task.create(
+          title: workflow_task.fetch("title"),
+          hint: workflow_task.fetch("hint", nil),
+          guidance_summary: workflow_task.fetch("guidance_summary", nil),
+          guidance_text: workflow_task.fetch("guidance_text", nil),
+          order: index,
+          section: section
+        )
 
         create_actions(workflow_task, task)
       end