+++ db/migrate/20220818142154_add_hints_and_guidance_to_tasks.rb
@@ -1,0 +1,7 @@
+class AddHintsAndGuidanceToTasks < ActiveRecord::Migration[7.0]
+  def change
+    add_column :tasks, :hint, :text, null: true
+    add_column :tasks, :guidance_summary, :string, null: true
+    add_column :tasks, :guidance_text, :text, null: true
+  end
+end